feat: Track user statistics for new user stats API
In a previous approach course stats were being generated on the fly for counts of user posts, responses and flags, however this would result in overly large amounts of data in the case of larger courses. This commit instead maintains these counts in the database so that this data can be fetched using a simpler, paginated and sortable query.